# No longer require external call to `initializeAnyploidPLIndexToAlleleIndices`