### Description
 
Instead of requiring anyploidy PLIndex caches to be initialized by an external call to `initializeAnyploidPLIndexToAlleleIndices`, calculate and cache PLIndex allele lists on an as needed basis.  Since `initializeAnyploidPLIndexToAlleleIndices` is public and `anyploidPloidyToPLIndexToAlleleIndices` is protected, they remain in the code to avoid breaking the API, but are deprecated.

### Things to think about before submitting:
- [ ] Make sure your changes compile and new tests pass locally.
- [ ] Add new tests or update existing ones:
  - A bug fix should include a test that previously would have failed and passes now.
  - New features should come with new tests that exercise and validate the new functionality.
- [ ] Extended the README / documentation, if necessary
- [ ] Check your code style.
- [ ] Write a clear commit title and message
  - The commit message should describe what changed and is targeted at htsjdk developers
  - Breaking changes should be mentioned in the commit message.