@kachulis 2020-08-18
Some clarification on the justification for this PR.  While the PR was motivated as part of a performance improvement in gatk (broadinstitute/gatk#6754), I believe it is a reasonable change to htsjdk in its own right as well.

Currently (before this PR) in order to use the static method `GenotypeLikelihoods.getAlleles` with any ploidy other than 2, any calling code must first call `GenotypeLikelihoods.initializeAnyploidPLIndexToAlleleIndices`.  This increases the complexity of using GenotypeLikelihoods in non ploidy-2 situations.   Removing this additional complexity of use without sacrificing performance is the objective of this PR.