@kachulis 2020-09-02
Thanks for the review @lbergelson!

Regarding performance, as far as I can tell, this code is essentially unused in gatk, picard, or the rest of htsjdk (save for the one issue referred to in broadinstitute/gatk#6754, which only uses code removed by this pr).  The only place in either gatk, picard, or htsjdk that uses `GenotypeLikelihoods::getAlleles` is in picard's `AlleleSubsettingUtils::subsetAlleles`, which is only implemented for ploidy 2 currently, so doesn't hit any of the anyPloidy code.  GATK appears to generally use its own `GenotypeLikelihoodCalculator` class instead of `GenotypeLikelihoods`.  So, to confirm performance, I wrote up a quick and hacky performance test (`GenotypeLikelihoodsUnitTest::performanceTest` in the `ck_anyploidy_new_performance_test` and `ck_anyploidy_new_performance_test` branches on my fork).  Ran with a few different ploidies, with results below.

ploidy 3:
old 3.52 s
new 2.02 s

ploidy 4:
old 2.74 s
new 3.37 s

ploidy 5:
old 3.18 s
new 2.08 s

which leads me to conclude that the performance is basically equivalent.

back to you!